Moved android platform impls into android library, added android library publish target
