+++ android/crt/build.gradle
@@ -4,8 +4,8 @@ apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
-
-preBuild.doFirst {
+// Before configuring, make sure libcrypto is installed
+preBuild {
     println('Installing libcrypto prebuilt binaries')
     def exec = (project.file('fetch_libcrypto.sh').absolutePath).execute()
     exec.waitForProcessOutput(System.out, System.err)
@@ -13,7 +13,6 @@ preBuild.doFirst {
 
 Properties getGitTag() {
     def gitTag = "git describe --tags".execute().text.trim()
-    println("gitTag=${gitTag}")
     def version = new Properties()
     def versionPattern = Pattern.compile('v(\\d+).(\\d+).(\\d+)(-(.+))?')
     def matcher = versionPattern.matcher(gitTag)
@@ -25,7 +24,6 @@ Properties getGitTag() {
             version['tag'] = matcher.group(5)
         }
     }
-    println("VERSION=${version}")
     return version
 }
 
@@ -33,7 +31,6 @@ ext {
     gitVersionName = {
         def version = getGitTag()
         def name = "${version['major']}.${version['minor']}.${version['patch']}"
-        println("versionName=${name}")
         return name
     }
     gitVersionCode = {
@@ -42,7 +39,6 @@ ext {
             def major = version['major'] as int
             def minor = version['minor'] as int
             def patch = version['patch'] as int
-            println("major=${major}, minor=${minor}, patch=${patch}")
             return (major * 1000) + (minor * 100) + patch
         } catch (Exception ex) {
             return 0
@@ -93,7 +89,7 @@ android {
             setRoot '../../src/test'
             java.srcDir '../../src/test/java'
             java.srcDir 'src/androidTest/java'
-            assets.srcDir 'src/androidTest/res'
+            assets.srcDir 'src/androidTest/assets'
         }
     }
 
@@ -131,4 +127,63 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test:monitor:1.2.0@aar'
     androidTestImplementation 'androidx.test:rules:1.2.0'
+}
+
+// Publishing
+apply plugin: 'maven-publish'
+
+// Sources
+task androidSourcesJar(type: Jar) {
+    archiveClassifier.set('sources')
+    from android.sourceSets.main.java.srcDirs
+}
+
+// Docs
+task androidDocs(type: Javadoc) {
+    source = android.sourceSets.main.java.srcDirs
+    classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
+    android.libraryVariants.all { variant ->
+        if (variant.name == 'release') {
+            owner.classpath += variant.javaCompileProvider.get().classpath
+        }
+    }
+    exclude '**/R.html', '**/R.*.html', '**/index.html'
+}
+
+task androidDocsJar(type: Jar) {
+    archiveClassifier.set('javadoc')
+    from androidDocs.destinationDir
+}
+
+afterEvaluate {
+    publishing {
+        publications {
+            // Creates a Maven publication called "release".
+            release(MavenPublication) {
+                // Applies the component for the release build variant.
+                from components.release
+
+                // You can then customize attributes of the publication as shown below.
+                groupId = 'software.amazon.awssdk.crt'
+                artifactId = 'android'
+                version = android.defaultConfig.versionName
+            }
+            // Creates a Maven publication called “debug”.
+            debug(MavenPublication) {
+                // Applies the component for the debug build variant.
+                from components.debug
+
+                groupId = 'software.amazon.awssdk.crt'
+                artifactId = 'android'
+                version = android.defaultConfig.versionName + '-SNAPSHOT'
+            }
+        }
+        repositories {
+            maven {
+                def snapshotRepo = "https://aws.oss.sonatype.org/content/repositories/snapshots"
+                def releaseRepo = "https://aws.oss.sonatype.org/"
+                url = version.endsWith('SNAPSHOT') ? snapshotRepo : releaseRepo
+            }
+        }
+    }
 }
